requirements.txt
requests==2.28.1
langchain==0.17.0
langchain-openai==0.0.9
composio-langchain==0.1.0
python-dotenv==1.0.0
langchain-core==0.1.0
